requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
bleach==4.1.0
bootstrap4==0.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.47.0
django-bootstrap4==21.2
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-embed-video==1.4.1
django-filter==21.1
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-storages==1.12.3
django-summernote==0.8.20.0
django-taggit==1.5.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
packaging==21.3
Pillow==9.0.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2020.4
requests==2.27.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.3.1
sqlparse==0.4.1
urllib3==1.26.8
webencodings==0.5.1
whitenoise==5.3.0
